requirements.txt
# ResumeAI backend
anthropic>=0.39.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pypdf>=5.0.0
python-docx>=1.0.0
